+++ .gitignore
@@ -1,2 +1,18 @@
-*.log
+# See http://help.github.com/ignore-files/ for more about ignoring files.
+
+# dependencies
 node_modules
+
+# testing
+coverage
+
+# production
+build
+
+# misc
+.DS_Store
+.env
+npm-debug.log
+
+
+*.log
